Refactored FXQuote,GetFXQuote to accept Amount in minor

+++ nanopay/src/net/nanopay/fx/FXService.js
@@ -18,7 +18,7 @@ foam.INTERFACE({
                 },
                 {
                     name: 'sourceAmount',
-                    javaType: 'double'
+                    javaType: 'Long'
                 },
                 {
                     class: 'String',
